Create cc_markov.py
This file creates a class called MarkovChain that allows a user to define a list as a Markov chain and then prompt the code to generate a response to a user input according to that chain. In this way, the system will respond in a way similar to that of the database.